Refactor coreutils token
Might not make sense, given the prevalence of coreutils imports
